Crate rusoto_resource_groups

AWS Resource Groups
AWS Resource Groups lets you organize AWS resources such as Amazon EC2 instances, Amazon Relational Database Service databases, and Amazon S3 buckets into groups using criteria that you define as tags. A resource group is a collection of resources that match the resource types specified in a query, and share one or more tags or portions of tags. You can create a group of resources based on their roles in your cloud infrastructure, lifecycle stages, regions, application layers, or virtually any criteria. Resource groups enable you to automate management tasks, such as those in AWS Systems Manager Automation documents, on tag-related resources in AWS Systems Manager. Groups of tagged resources also let you quickly view a custom console in AWS Systems Manager that shows AWS Config compliance and other monitoring data about member resources.

To create a resource group, build a resource query, and specify tags that identify the criteria that members of the group have in common. Tags are key-value pairs.

For more information about Resource Groups, see the AWS Resource Groups User Guide.

AWS Resource Groups uses a REST-compliant API that you can use to perform the following types of operations.

Create, Read, Update, and Delete (CRUD) operations on resource groups and resource query entities

Applying, editing, and removing tags from resource groups

Resolving resource group member ARNs so they can be returned as search results

Getting data about resources that are members of a group

Searching AWS resources based on a resource query

If you're using the service, you're probably looking for ResourceGroupsClient and ResourceGroups.

Structs

CreateGroupInput

CreateGroupOutput

DeleteGroupInput

DeleteGroupOutput

GetGroupInput

GetGroupOutput

GetGroupQueryInput

GetGroupQueryOutput

GetTagsInput

GetTagsOutput

Group
A resource group.

GroupFilter
A filter name and value pair that is used to obtain more specific results from a list of groups.

GroupIdentifier
The ARN and group name of a group.

GroupQuery
The underlying resource query of a resource group. Resources that match query results are part of the group.

ListGroupResourcesInput

ListGroupResourcesOutput

ListGroupsInput

ListGroupsOutput

QueryError
A two-part error structure that can occur in ListGroupResources or SearchResources operations on CloudFormation stack-based queries. The error occurs if the CloudFormation stack on which the query is based either does not exist, or has a status that renders the stack inactive. A QueryError occurrence does not necessarily mean that AWS Resource Groups could not complete the operation, but the resulting group might have no member resources.

ResourceFilter
A filter name and value pair that is used to obtain more specific results from a list of resources.

ResourceGroupsClient
A client for the Resource Groups API.

ResourceIdentifier
The ARN of a resource, and its resource type.

ResourceQuery
The query that is used to define a resource group or a search for resources.

SearchResourcesInput

SearchResourcesOutput

TagInput

TagOutput

UntagInput

UntagOutput

UpdateGroupInput

UpdateGroupOutput

UpdateGroupQueryInput

UpdateGroupQueryOutput

Enums

CreateGroupError
Errors returned by CreateGroup

DeleteGroupError
Errors returned by DeleteGroup

GetGroupError
Errors returned by GetGroup

GetGroupQueryError
Errors returned by GetGroupQuery

GetTagsError
Errors returned by GetTags

ListGroupResourcesError
Errors returned by ListGroupResources

ListGroupsError
Errors returned by ListGroups

SearchResourcesError
Errors returned by SearchResources

TagError
Errors returned by Tag

UntagError
Errors returned by Untag

UpdateGroupError
Errors returned by UpdateGroup

UpdateGroupQueryError
Errors returned by UpdateGroupQuery

Traits

ResourceGroups
Trait representing the capabilities of the Resource Groups API. Resource Groups clients implement this trait.