Struct rusoto_resource_groups::ResourceGroupsClient

pub struct ResourceGroupsClient { /* fields omitted */ }

A client for the Resource Groups API.

Methods

impl ResourceGroupsClient

pub fn new(region: Region) -> ResourceGroupsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ResourceGroupsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,