Trait rusoto_resource_groups::ResourceGroups

pub trait ResourceGroups {
    fn create_group(
        &self, 
        input: CreateGroupInput
    ) -> RusotoFuture<CreateGroupOutput, CreateGroupError>;

    fn delete_group(
        &self, 
        input: DeleteGroupInput
    ) -> RusotoFuture<DeleteGroupOutput, DeleteGroupError>;

    fn get_group(
        &self, 
        input: GetGroupInput
    ) -> RusotoFuture<GetGroupOutput, GetGroupError>;

    fn get_group_query(
        &self, 
        input: GetGroupQueryInput
    ) -> RusotoFuture<GetGroupQueryOutput, GetGroupQueryError>;

    fn get_tags(
        &self, 
        input: GetTagsInput
    ) -> RusotoFuture<GetTagsOutput, GetTagsError>;

    fn list_group_resources(
        &self, 
        input: ListGroupResourcesInput
    ) -> RusotoFuture<ListGroupResourcesOutput, ListGroupResourcesError>;

    fn list_groups(
        &self, 
        input: ListGroupsInput
    ) -> RusotoFuture<ListGroupsOutput, ListGroupsError>;

    fn search_resources(
        &self, 
        input: SearchResourcesInput
    ) -> RusotoFuture<SearchResourcesOutput, SearchResourcesError>;

    fn tag(&self, input: TagInput) -> RusotoFuture<TagOutput, TagError>;

    fn untag(&self, input: UntagInput) -> RusotoFuture<UntagOutput, UntagError>;

    fn update_group(
        &self, 
        input: UpdateGroupInput
    ) -> RusotoFuture<UpdateGroupOutput, UpdateGroupError>;

    fn update_group_query(
        &self, 
        input: UpdateGroupQueryInput
    ) -> RusotoFuture<UpdateGroupQueryOutput, UpdateGroupQueryError>;
}

Trait representing the capabilities of the Resource Groups API. Resource Groups clients implement this trait.

Required methods

fn create_group(
&self,
input: CreateGroupInput
) -> RusotoFuture<CreateGroupOutput, CreateGroupError>

Creates a group with a specified name, description, and resource query.

fn delete_group(
&self,
input: DeleteGroupInput
) -> RusotoFuture<DeleteGroupOutput, DeleteGroupError>

Deletes a specified resource group. Deleting a resource group does not delete resources that are members of the group; it only deletes the group structure.

fn get_group(
&self,
input: GetGroupInput
) -> RusotoFuture<GetGroupOutput, GetGroupError>

Returns information about a specified resource group.

fn get_group_query(
&self,
input: GetGroupQueryInput
) -> RusotoFuture<GetGroupQueryOutput, GetGroupQueryError>

Returns the resource query associated with the specified resource group.

fn get_tags(
&self,
input: GetTagsInput
) -> RusotoFuture<GetTagsOutput, GetTagsError>

Returns a list of tags that are associated with a resource group, specified by an ARN.

fn list_group_resources(
&self,
input: ListGroupResourcesInput
) -> RusotoFuture<ListGroupResourcesOutput, ListGroupResourcesError>

Returns a list of ARNs of resources that are members of a specified resource group.

fn list_groups(
&self,
input: ListGroupsInput
) -> RusotoFuture<ListGroupsOutput, ListGroupsError>

Returns a list of existing resource groups in your account.

fn search_resources(
&self,
input: SearchResourcesInput
) -> RusotoFuture<SearchResourcesOutput, SearchResourcesError>

Returns a list of AWS resource identifiers that matches a specified query. The query uses the same format as a resource query in a CreateGroup or UpdateGroupQuery operation.

fn tag(&self, input: TagInput) -> RusotoFuture<TagOutput, TagError>

Adds tags to a resource group with the specified ARN. Existing tags on a resource group are not changed if they are not specified in the request parameters.

fn untag(&self, input: UntagInput) -> RusotoFuture<UntagOutput, UntagError>

Deletes specified tags from a specified resource.

fn update_group(
&self,
input: UpdateGroupInput
) -> RusotoFuture<UpdateGroupOutput, UpdateGroupError>

Updates an existing group with a new or changed description. You cannot update the name of a resource group.

fn update_group_query(
&self,
input: UpdateGroupQueryInput
) -> RusotoFuture<UpdateGroupQueryOutput, UpdateGroupQueryError>

Updates the resource query of a group.

Loading content...